# perf: module graph url shortcuts