--- a/packages/vite/src/node/server/moduleGraph.ts
+++ b/packages/vite/src/node/server/moduleGraph.ts
@@ -265,22 +265,17 @@ export class ModuleGraph {
     url = removeImportQuery(removeTimestampQuery(url))
     const resolved = await this.resolveId(url, !!ssr)
     const resolvedId = resolved?.id || url
-    url = ensureExtension(url, resolvedId)
+    if (
+      url !== resolvedId &&
+      !url.includes('\0') &&
+      !url.startsWith(`virtual:`)
+    ) {
+      const ext = extname(cleanUrl(resolvedId))
+      const { pathname, search, hash } = new URL(url, 'relative://')
+      if (ext && !pathname!.endsWith(ext)) {
+        url = pathname + ext + search + hash
+      }
+    }
     return [url, resolvedId, resolved?.meta]
   }
-}
-
-function ensureExtension(url: string, resolvedId: string) {
-  if (
-    url !== resolvedId &&
-    !url.includes('\0') &&
-    !url.startsWith(`virtual:`)
-  ) {
-    const ext = extname(cleanUrl(resolvedId))
-    const { pathname, search, hash } = new URL(url, 'relative://')
-    if (ext && !pathname!.endsWith(ext)) {
-      return pathname + ext + search + hash
-    }
-  }
-  return url
 }