--- a/packages/vite/src/node/server/transformRequest.ts
+++ b/packages/vite/src/node/server/transformRequest.ts
@@ -144,7 +144,9 @@ async function doTransform(
 
   // resolve
   const id =
-    (await pluginContainer.resolveId(url, undefined, { ssr }))?.id || url
+    module?.id ??
+    (await pluginContainer.resolveId(url, undefined, { ssr }))?.id ??
+    url
 
   const result = loadAndTransform(id, url, server, options, timestamp)