### Description

Implement shortcuts based on [this idea](https://github.com/vitejs/vite/issues/12363#issuecomment-1487304874) by @sapphi-red

It looks like there are opportunities here. But I don't see an improvement in the tests we were running. Sending the PR as an exploration, maybe a similar one should be merged if we can prove it helps.

---

### What is the purpose of this pull request? <!-- (put an "X" next to an item) -->

- [ ] Bug fix
- [ ] New Feature
- [ ] Documentation update
- [x] Other