@sun0day 2023-03-29
I tried to cache the `ensureEntryFromUrl` result before. But it doesn't seem to help the performance much.