@sapphi-red 2023-03-29
Thanks!

The result of running https://github.com/sapphi-red/performance-compare/tree/65d51a32b0282593f1b75def25075a02b28f6397 on my machine.

### Without this PR
```
Vite  startup time: 4830ms (including server start up time: 511ms)
Vite  Root HMR time: 36ms
Vite  Leaf HMR time: 16ms
Vite (swc)  startup time: 3534ms (including server start up time: 243ms)
Vite (swc)  Root HMR time: 33ms
Vite (swc)  Leaf HMR time: 12ms
```

### With this PR
```
Vite  startup time: 4300ms (including server start up time: 481ms)
Vite  Root HMR time: 34ms
Vite  Leaf HMR time: 15ms
Vite (swc)  startup time: 2324ms (including server start up time: 247ms)
Vite (swc)  Root HMR time: 29ms
Vite (swc)  Leaf HMR time: 11ms
```