@sapphi-red 2023-03-29
> @sapphi-red maybe you could test with https://github.com/sun0day/vite-2.7-slow on your computer? And do you have insights into why is the difference?

This PR improves performance of the non-pre-bundled files. Because that repo only includes a few project files, this PR won't have any impact.

> Is it still that even with realpath.native, it is still a bottleneck for Windows? Do you have a CPU profile to share?

I think it's still a bottleneck, but not as bad as before.
[cpuprofiles.zip](https://github.com/vitejs/vite/files/11097835/cpuprofiles.zip) (It seems you can't upload the `.cpuprofile` files directly.)