@sapphi-red 2023-03-29
I forgot to pass `--force` when taking the previous cpuprofiles. 😅

This is the correct one. This time, I took with SWC one too.
[cpuprofiles-new.zip](https://github.com/vitejs/vite/files/11098001/cpuprofiles-new.zip)