@patak-dev 2023-03-29
@sapphi-red would it be possible to run the tests again on your machine to see if the speed-up was due to the current changes in the PR or by the other optimization that I removed (and if that helps I can send in a separate PR, but it was generating an error in vite-plugin-vue for some reason)