@sapphi-red 2023-03-29
Here's the result 🙂
```
Vite  startup time: 4378ms (including server start up time: 487ms)
Vite  Root HMR time: 34ms
Vite  Leaf HMR time: 14ms
Vite (swc)  startup time: 2410ms (including server start up time: 240ms)
Vite (swc)  Root HMR time: 29ms
Vite (swc)  Leaf HMR time: 12ms
```