@patak-dev 2023-03-29
Thanks @sapphi-red! So it seems that the biggest perf boost is from the rawUrl cache, but the resolving short-circuit also helps (I'll send another PR and we can check there why plugin-vue was failing).

@bluwy moved the cache to a new `#rawUrlToModuleMap`, and made it private so we don't need to yet expose this as part of the ModuleGraph API. I think we could merge this one.