@patak-dev 2023-03-29
Thanks for fixing tsconf.check.json @ArnaudBarre! @sapphi-red could you check this looks good to you too?