@patak-dev 2023-03-30
Converting this to a draft until we better understand the implications of this cache. @bluwy, I think that `rawUrlToModule` needs to also depend on `ssr`. resolveId for a certain url can end up in different ids (and resolved urls).

@ArnaudBarre would you send a PR with the changes to `tsconfig.check.js` in a separate PR so we can discuss and merge that one?