@patak-dev 2023-03-30
@ArnaudBarre @bluwy @sapphi-red looks like using private identifiers is making other CIs fail. From the SvelteKit one:
```
  Error: packages/adapter-cloudflare-workers check: ../../node_modules/.pnpm/file+..+..+vite+packages+vite_@types+node@16.18.6/node_modules/vite/dist/node/index.d.ts(1318,5): error TS18028: Private identifiers are only available when targeting ECMAScript 2015 and higher.
```

Should we try to push for everyone to update to ES2020? Maybe I should revert to using WeakMaps for now and we can push this change in Vite 5.

Still is strange because we already have ES2020 in our `tsconfig.json`. The diff couldn't introduce subtle bugs?