@sapphi-red 2023-03-30
> Thanks for fixing tsconf.check.json @ArnaudBarre! @sapphi-red could you check this looks good to you too?

Yeah, it looks good to me 👍