@patak-dev 2023-03-30
It looks like we're also doing a double `resolve` of the URL here: [vitejs/vite@`211df5d` (#12635)](https://github.com/vitejs/vite/pull/12635/commits/211df5d45d929dd431ba6b7ddd665e6b049b05e2)

I think we don't notice these as much because the `url` incoming in `transformRequest` has been already resolved and is mostly an absolute-looking path starting from root already with the proper extension. There will still be a realpath call though every time we call `resolveUrl` before this PR.