@patak-dev 2023-03-30
Same projects are failing in main https://github.com/vitejs/vite-ecosystem-ci/actions/runs/4567006041, probably due to changes to vite-plugin-vue today. I think we can merge this one to test it during the beta